requirements.txt
Flask==3.0.0
groq==0.4.1
PyPDF2==3.0.1
python-docx==1.1.0
Werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
python-pptx==0.6.23
wordcloud==1.9.3
Pillow==10.1.0
numpy==1.26.2
pdfkit==1.0.0
matplotlib==3.8.2
